Unsigned types changes
In Java there is no unsigned types for int, byte, short, long. So if we map this types to same with signed, artihmetic overflow exception possibility come out. So I think in here map will be a upper range. For example byte in C# maps to short in Java, go on like that until ulong in C# maps to BigInteger.

Another missing case here, .Net 6 will go out with DateOnly and TimeOnly fields. If Hazelcast thinks to support .Net 6, this types will be here.