Deal with Building Warnings- Part 3 [API-2075] [CTT-744] (#994)
Prepared on top of
https://github.com/hazelcast/hazelcast-csharp-client/pull/993

The high number of changes come from public API definitions in
`PublicAPI` folder. Although public APIs are already defined for .NET8
for example in those text files, `NET9`, `netstandart2.0` and
`netstandart2.1` targets were missing. So the content is same. The
compiler was giving warnings due to missing definitions on those target
builds. Compiler uses that files to warn on new public APIs. The APIs
moved to `shipped.txt` file are ignored.